unify shared-component snapshot regen into the auto-update workflow
`generate-expected-html.ts` already runs on every PR that touches the
compiler / Hono / client / adapter-tests, regenerates the inline
`expectedHtml` for the conformance corpus, and commits any drift back
to the PR.

Shared-component fixtures (fixture-hydrate corpus) write their
`expectedHtml` + `expectedClientJs` to
`fixtures/__snapshots__/<id>.{html,client.js}` instead of inline
strings — so the existing generator never touched them. Adding a
parallel workflow would duplicate the trigger / commit boilerplate;
slotting a second step into the existing job would split the
generation across two scripts.

Lift the snapshot pipeline out of `scripts/snapshot.ts` into
`src/snapshot-generator.ts` (a library module both scripts call),
and run `generateAllSharedComponentSnapshots()` at the tail of
`generate-expected-html.ts`. The CLI `scripts/snapshot.ts` shrinks
to a thin wrapper that exists for interactive `bun run` invocations.

### Workflow trigger expansion

`update-fixtures.yml` now also fires on changes to
`integrations/shared/components/**` (the .tsx sources whose output the
snapshots pin) and `packages/adapter-tests/fixtures/**` (new specs /
prop changes / additionalComponents edits). The commit step already
stages `packages/adapter-tests/fixtures/` recursively, so the new
`__snapshots__/` deltas ride along without further workflow changes.

The "Check for changes" diff gate keeps efficiency intact — clean runs
on irrelevant compiler changes complete without committing.
