+++ .github/workflows/update-fixtures.yml
@@ -9,6 +9,13 @@ on:
       - 'packages/client/**'
       - 'packages/adapter-tests/src/**'
       - 'packages/adapter-tests/scripts/**'
+      # Shared-component fixture sources (fixture-hydrate corpus inputs).
+      # Editing a `*.tsx` here regenerates the `__snapshots__/<id>.{html,client.js}` pair.
+      - 'integrations/shared/components/**'
+      # Fixture spec files (adding a new fixture, changing its props, or
+      # adjusting `componentName` / `sourceFile` / `additionalComponents`
+      # changes the snapshot output).
+      - 'packages/adapter-tests/fixtures/**'
 
 permissions:
   contents: write
